+++ processor/src/test/fixtures/expected/com/example/prop/ImmutablePropTypeDynamicValueBuilder.java
@@ -1,0 +1,48 @@
+package com.example.prop;
+
+import arez.component.Identifiable;
+import javax.annotation.Generated;
+import javax.annotation.Nonnull;
+import react4j.Keyed;
+import react4j.React;
+import react4j.ReactElement;
+import react4j.ReactNode;
+
+@Generated("react4j.processor.React4jProcessor")
+final class ImmutablePropTypeDynamicValueBuilder {
+  private ImmutablePropTypeDynamicValueBuilder() {
+  }
+
+  @Nonnull
+  private static Step1 newBuilder() {
+    return new Builder();
+  }
+
+  @Nonnull
+  static ReactNode myProp(final Object myProp) {
+    return newBuilder().myProp( myProp );
+  }
+
+  public interface Step1 {
+    @Nonnull
+    ReactNode myProp(Object myProp);
+  }
+
+  private static class Builder implements Step1 {
+    private final ReactElement _element = ReactElement.createViewElement( React4j_ImmutablePropTypeDynamicValue.Factory.TYPE );
+
+    @Override
+    @Nonnull
+    public final ReactNode myProp(final Object myProp) {
+      _element.setKey( ( myProp instanceof Keyed ? Keyed.getKey( myProp ) : myProp instanceof Identifiable ? Identifiable.<Object>getArezId( myProp ) : String.valueOf( myProp ) ) + ( React.enableViewNames() ? "_ImmutablePropTypeDynamicValue_a98f656b" : ImmutablePropTypeDynamicValue.class.getName() ) );
+      _element.input( React4j_ImmutablePropTypeDynamicValue.Inputs.myProp, myProp );
+      return build();
+    }
+
+    @Nonnull
+    public final ReactNode build() {
+      _element.complete();
+      return _element;
+    }
+  }
+}