+++ processor/src/test/fixtures/expected/com/example/prop/ImmutablePropTypesBuilder.java
@@ -46,10 +46,15 @@ final class ImmutablePropTypesBuilder {
 
   public interface Step5 {
     @Nonnull
-    ReactNode someProp(ImmutablePropTypes.Foo someProp);
+    Step6 someProp(ImmutablePropTypes.Foo someProp);
   }
 
-  private static class Builder implements Step1, Step2, Step3, Step4, Step5 {
+  public interface Step6 {
+    @Nonnull
+    ReactNode object(Object object);
+  }
+
+  private static class Builder implements Step1, Step2, Step3, Step4, Step5, Step6 {
     private final ReactElement _element = ReactElement.createViewElement( React4j_ImmutablePropTypes.Factory.TYPE );
 
     @Override
@@ -82,15 +87,23 @@ final class ImmutablePropTypesBuilder {
 
     @Override
     @Nonnull
-    public final ReactNode someProp(final ImmutablePropTypes.Foo someProp) {
+    public final Step6 someProp(final ImmutablePropTypes.Foo someProp) {
       _element.input( React4j_ImmutablePropTypes.Inputs.someProp, someProp );
+      return this;
+    }
+
+    @Override
+    @Nonnull
+    public final ReactNode object(final Object object) {
+      _element.input( React4j_ImmutablePropTypes.Inputs.object, object );
       return build();
     }
 
     @Nonnull
     public final ReactNode build() {
       final JsPropertyMap<Object> inputs = _element.inputs();
-      _element.setKey( String.valueOf( Identifiable.<Object>getArezId( (ImmutablePropTypes.MyComponent) inputs.get( React4j_ImmutablePropTypes.Inputs.myProp ) ) ) + "-" + ( (String) inputs.get( React4j_ImmutablePropTypes.Inputs.myOtherProp ) ) + "-" + String.valueOf( (int) (double) inputs.get( React4j_ImmutablePropTypes.Inputs.stillAnotherProp ) ) + "-" + Keyed.getKey( (ImmutablePropTypes.KeyedComponent) inputs.get( React4j_ImmutablePropTypes.Inputs.BobsProp ) ) + "-" + ( (ImmutablePropTypes.Foo) inputs.get( React4j_ImmutablePropTypes.Inputs.someProp ) ) + ( React.enableViewNames() ? "_ImmutablePropTypes_108bb2eb" : ImmutablePropTypes.class.getName() ) );
+      final Object $object$ = inputs.get( React4j_ImmutablePropTypes.Inputs.object );
+      _element.setKey( String.valueOf( Identifiable.<Object>getArezId( (ImmutablePropTypes.MyComponent) inputs.get( React4j_ImmutablePropTypes.Inputs.myProp ) ) ) + "-" + ( (String) inputs.get( React4j_ImmutablePropTypes.Inputs.myOtherProp ) ) + "-" + String.valueOf( (int) (double) inputs.get( React4j_ImmutablePropTypes.Inputs.stillAnotherProp ) ) + "-" + Keyed.getKey( (ImmutablePropTypes.KeyedComponent) inputs.get( React4j_ImmutablePropTypes.Inputs.BobsProp ) ) + "-" + ( (ImmutablePropTypes.Foo) inputs.get( React4j_ImmutablePropTypes.Inputs.someProp ) ) + "-" + ( $object$ instanceof Keyed ? Keyed.getKey( $object$ ) : $object$ instanceof Identifiable ? Identifiable.<Object>getArezId( $object$ ) : String.valueOf( $object$ ) ) + ( React.enableViewNames() ? "_ImmutablePropTypes_108bb2eb" : ImmutablePropTypes.class.getName() ) );
       _element.complete();
       return _element;
     }