+++ processor/src/test/fixtures/input/com/example/prop/ImmutablePropTypeDynamicValue.java
@@ -1,7 +1,5 @@
 package com.example.prop;
 
-import arez.annotations.ArezComponent;
-import arez.annotations.Feature;
 import javax.annotation.Nullable;
 import react4j.ReactNode;
 import react4j.annotations.Input;
@@ -9,15 +7,10 @@ import react4j.annotations.Render;
 import react4j.annotations.View;
 
 @View
-abstract class ImmutablePropTypeArezComponentHasNoId
+abstract class ImmutablePropTypeDynamicValue
 {
-  @ArezComponent( allowEmpty = true, requireId = Feature.DISABLE )
-  static abstract class MyComponent
-  {
-  }
-
   @Input( immutable = true )
-  protected abstract MyComponent getMyProp();
+  abstract Object getMyProp();
 
   @Nullable
   @Render