+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Fix bug where `STATELESS` views were not disposed unless they had reactive components. This resulted in views being cached forever in development builds and errors in production builds if callbacks used `Disposable.isDisposed(this)` to determine whether the view was still mounted.
 * Rework `textarea` host element factor in the dom module to stop support child elements as the underlying react.development.js library will generate a warning if a child component is present.
 * Add the `react4j-webspeechdemo` to the variants tracked as part of the release process.
+* Allow arbitrary types for immutable props. To generate the key for the input, the framework will inspect the value at runtime to detect if it implements `react4j.Keyed` and use that, otherwise use `Identifiable` if that is implemented by the value otherwise `String.valueOf()` is used to convert the value to key segment.
 
 ### [v0.168](https://github.com/react4j/react4j/tree/v0.168) (2020-06-25) · [Full Changelog](https://github.com/react4j/react4j/compare/v0.167...v0.168)